+++ lib/src/main/kotlin/databaseManage/RBTreeManager.kt
@@ -3,7 +3,7 @@ package databaseManage
 import databaseSave.neo4j.DrawableRBVertex
 import databaseSave.neo4j.Neo4jRepository
 import treelib.rbTree.RBStruct
-import treelib.singleObjects.Container
+import treelib.commonObjects.Container
 
 class RBTreeManager {
