+++ lib/src/main/kotlin/treelib/abstractTree/balanced/BalancerParent.kt
@@ -2,7 +2,7 @@ package treelib.abstractTree.balanced
 
 import treelib.abstractTree.NodeParent
 import treelib.abstractTree.StateContainer
-import treelib.singleObjects.exceptions.IllegalNodeStateException
+import treelib.commonObjects.exceptions.IllegalNodeStateException
 
 abstract class BalancerParent<Pack : Comparable<Pack>, NodeType : NodeParent<Pack, NodeType>, StateContainerType : StateContainer<Pack, NodeType>> :
     Balancer<Pack, NodeType, StateContainerType> {
