+++ lib/src/main/kotlin/treelib/avlTree/AVLTree.kt
@@ -2,7 +2,7 @@ package treelib.avlTree
 
 import treelib.abstractTree.Tree
 
-import treelib.singleObjects.Container
+import treelib.commonObjects.Container
 
 class AVLTree<K : Comparable<K>, V> :
     Tree<K, V, AVLNode<Container<K, V?>>, AVLStateContainer<Container<K, V?>>, AVLVertex<Container<K, V?>>>() {
