+++ lib/src/main/kotlin/treelib/commonObjects/Container.kt
@@ -1,8 +1,9 @@
-package treelib.singleObjects
+package treelib.commonObjects
 
 import kotlinx.serialization.Serializable
 
-class Container<K : Comparable<K>, V>(val pair: Pair<K, V?>) : Comparable<Container<K, V>> {
+@Serializable
+data class Container<K : Comparable<K>, V>(val pair: Pair<K, V?>) : Comparable<Container<K, V>> {
 
     val key = pair.first
     val value = pair.second
