+++ lib/src/main/kotlin/treelib/commonObjects/exceptions/IllegalBaseNodeException.kt
@@ -1,4 +1,4 @@
-package treelib.singleObjects.exceptions
+package treelib.commonObjects.exceptions
 
 class IllegalBaseNodeException : Exception {
     constructor() : super(
