+++ lib/src/main/kotlin/treelib/rbTree/Markers.kt
@@ -1,4 +1,4 @@
-package treelib.singleObjects
+package treelib.rbTree
 
 enum class Markers {
     RED, BLACK
