+++ lib/src/main/kotlin/treelib/rbTree/RBVertex.kt
@@ -1,7 +1,6 @@
 package treelib.rbTree
 
 import treelib.abstractTree.Vertex
-import treelib.singleObjects.Markers
 
 open class RBVertex<Pack : Comparable<Pack>>(
     override val value: Pack,
