+++ lib/src/test/kotlin/treelib/AVLStructTest.kt
