+++ lib/src/test/kotlin/treelib/BINStructTest.kt
