+++ lib/src/test/kotlin/utils/TreeWrapper.kt
@@ -1,7 +1,7 @@
 package utils
 
 import treelib.abstractTree.*
-import treelib.singleObjects.Container
+import treelib.commonObjects.Container
 
 
 class TreeWrapper<
